requirements.txt
PyYAML==6.0
paramiko==2.11.0
termcolor==1.1.0